requirements.txt
# LangChain + LangGraph
langsmith>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Optional LLM providers (uncomment as needed)
# langchain-anthropic>=0.3.0
# langchain-ollama>=0.3.0

# Claim detector pre-filter
onnxruntime>=1.17.0
tokenizers>=0.19.0

# Search
# tavily-python>=0.5.0 # (uncomment if TAVILY_USE=true)
ddgs>=7.0.0

# Web content extraction
trafilatura>=1.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Domain analysis
python-whois>=0.9.5

# Backend
fastapi>=0.115.12
uvicorn[standard]>=0.30.0

# HTTP client
httpx>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Embeddings & similarity search
openai>=1.0.0
numpy>=1.24.0

# Data & config
pydantic>=2.0.0
json-repair>=0.30.0
python-dotenv>=1.0.0
rich>=13.0.0

# Tests
pytest>=8.0.0
